refactor(action & source): extract action and source list
